New GPIO Library

+++ src/comp/cylinder_triangle.rs
@@ -84,7 +84,7 @@ impl MathActor for CylinderTriangle
 
 impl SimpleMeas for CylinderTriangle
 {
-    fn init_meas(&mut self, pin_meas : u16) {
+    fn init_meas(&mut self, pin_meas : u8) {
         self.cylinder.init_meas(pin_meas)
     }
 }
