+++ src/meas.rs
@@ -1,4 +1,4 @@
 pub trait SimpleMeas 
 {
-    fn init_meas(&mut self, pin_meas : u16);
+    fn init_meas(&mut self, pin_meas : u8);
 }
